Add cluster option functionality
Cluster profiles can now make use of cluster options. Options are specified via
--cluster-option (-c) on the commandline and take the form of
profile.option=value to mimic the plugin options of sosreport.

Cluster profiles will have access to options that begin with the cluster type's
name, and no others. This can be particularly useful when needing to limit the
selection of nodes according to a cluster-specific reference.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>