Improve profiles ability to modify sosreport
Initial version of allowing profiles to modify the sosreport command being run,
or to modify the path to the sosreport in the case of Atomic Host, was ugly and
not very clean.

This adds methods to the profile class to make those modification easier when
writing cluster profiles.

This commit also updates the kubernetes profile to make use of these changes.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>