Simplify how profiles are enabled
This simplifies the way profiles are enabled. Instead of writing a
check_enabled() method for each profile, the profiles now just need to set a
'packages' tuple with the names of the package(s) the profile should be enabled
for.

Also corrected the kubernetes profile to use the correct command for OpenShift
clusters.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>