Improve handling of options handed to sosreport on nodes
Previously clustersos would hand individual options to sosreport individually,
menaing we could run sosreport with '-e plug1 -e plug2'. While this works, it is
not clean and was prone to breaking with how those options would collide when a
user specified that option on the commandline and a profile used it as well.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>