Add ovirt profile, change how extra data is collected
This adds an ovirt profile for situations in which log-collector is either
unavailable or unusable.

Additionally, since the ovirt profile is the first to make use of
run_extra_cmd() in a profile, this changes the mechanism by which it is called.

Minor corrections to commandline options for better parsing.

Minor fix involving name and case id not being applied to sosreports from nodes.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>