Enable using a non-root ssh user and leveraging sudo
This enables the changing of the ssh user to a non-root user who must then
use sudo on the remote nodes.

We only prompt for a sudo password once, assuming it is consistent across the
nodes. Users with different sudo passwords will need to either revert back to using
ssh keys for root access or sync their passwords. Prompting for multiple sudo passwords
would be clunky, inefficient and become kludgy very shortly.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>