Fix check for password line in exec_master_cmd
In exec_master_cmd(), there is a chance (and a certainty with python3) that the
returned output from splitlines would be byte-objects instead of strings. This
would cause the check removing the password prompt to raise an exception that
would break almost all commands run on a master node. By decoding the output to
strings before this check, instead of in the return, the check remains valid.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>