Correct duplicate entries when using shortnames and FQDNs
Previously, if a cluster was using shortnames for node names, and the localhost
was using an FQDN (or vice-versa) and was also a member of the cluster,
clustersos would try to collect a local sosreport and additionally try to
connect to the localhost via an ssh session to itself.

This is now corrected, and in the case of a mixed use of shortnames and FQDNs,
clustersos will revert to using the shortname internally, but still only run the
sosreport locally.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>