Make Atomic Host checks the default
This moves the check for if a node is an Atomic Host deployment
from the kubernetes and openshift profiles, making it the default
behavior.

If a profile needs to set prefix or cleanup commands now, it will
need to consider if the cluster type could possibly have Atomic Hosts
as nodes.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>